AccountManagerPlugin: Fill third wizard page with preliminary content, refs #8930.
Adding/correcting some options in other pages and creating a 6th one.

This turns out to become an interesting journey through the option jungle
grown by years of development for this plugin.  Hopefully the wizard will
become a reference showing not only, if/how every !AccountManager feature
is related to another one, but uncovering the big picture of Trac
authentication in general too.

git-svn-id: https://trac-hacks.org/svn/accountmanagerplugin@12570 7322e99d-02ea-0310-aa39-e9a107903beb